+++ CHANGELOG.md
@@ -13,11 +13,17 @@ Changelog](http://keepachangelog.com/).
 * [#117](https://gitlab.com/gitlab-data/permifrost/-/issues/117)
 Adds support for suffix partial schema name
 
+* [#116](https://gitlab.com/gitlab-data/permifrost/-/issues/116)
+Adds support for integrations
+
 ### Changes
 [#112](https://gitlab.com/gitlab-data/permifrost/-/issues/112)  Add support for 'grant select on all tables in db/schema'
 for `db.*.*` or `db.schema.*` in spec file.
 
 ### Fixes
+* [#125](https://gitlab.com/gitlab-data/permifrost/-/issues/125)
+Skip generation of write privileges for already-granted <database_or_schema>.* privileges
+
 * [#117](https://gitlab.com/gitlab-data/permifrost/-/issues/117)
 Remove `information_schema` of fetched schemas in case of partial schema name input